appointments

+++ db/migrate/20160820035027_create_doctor_prices.rb
@@ -3,7 +3,7 @@ class CreateDoctorPrices < ActiveRecord::Migration
     create_table :doctor_prices do |t|
       t.integer :doctor_id
       t.integer :clinic_id
-      t.integer :price
+      t.double :price
       t.timestamps null: false
     end
   end
